requirements.txt
streamlit==1.28.0
langchain==0.0.350
openai>=1.6.1,<2.0.0
faiss-cpu==1.7.4
pdfplumber==0.10.0
python-docx==0.8.11
python-dotenv==1.0.0
tiktoken==0.5.0
langchain-openai==0.0.2